Remove Nominations from Reasons
Nominations have no impact on whether or not something is on this list; only albums that won a grammy/AMA award are listed. at one point I was planning to add any album that was nominated to my list of albums to listen to, but I gave up on that pretty quickly.